menambahkan fitur filter dan search serta date time

+++ index.js
@@ -1,11 +1,20 @@
 const express = require('express')
+const fs = require('fs')
 const app = express()
 const port = 5000
 const expressLayout = require('express-ejs-layouts')
 const cors = require('cors')
-const body = require('body-parser')
+const bodyParser = require('body-parser')
 const multer = require('multer')
 const path = require('path')
+const {
+    body,
+    check,
+    validationResult
+} = require('express-validator');
+const {
+    Op
+} = require('sequelize');
 
 
 // upload image
@@ -18,11 +27,15 @@ const storage = multer.diskStorage({
         cb(null, Date.now() + '-' + file.originalname)
     }
 })
-
 const upload = multer({
     storage: storage
 })
 
+
+// manggil controller
+// const updatedAt = require('./controllers/CarController')
+
+// manggil model
 const {
     sequelize,
     connectToDB
@@ -30,13 +43,17 @@ const {
 
 const Car = require('./models/CarModels')
 
+
 app.set('view engine', 'ejs')
 
 app.use(cors())
 app.use(expressLayout)
 app.use(express.json())
 app.use(express.static(__dirname + '/public'))
-app.use(express.urlencoded())
+app.use(express.urlencoded({
+    extented: true
+}))
+
 
 
 app.get('/', (req, res) => {
@@ -48,14 +65,56 @@ app.get('/', (req, res) => {
 
 app.get('/cars', async (req, res) => {
 
-    const cars = await Car.findAll()
-    // res.status(200).json(cars)
+    const date = {
+        year: 'numeric',
+        month: 'long',
+        day: 'numeric',
+        hour: 'numeric',
+        minute: 'numeric',
+        timeZone: 'UTC'
+    };
+
+    if (req.query.filter) {
+        cars = await Car.findAll({
+            where: {
+                ukuran: {
+                    [Op.substring]: req.query.filter
+                }
+            },
+            order: [
+                ['updatedAt', 'DESC']
+            ]
+        })
+    } else {
+        cars = await Car.findAll({
+            order: [
+                ['updatedAt', 'DESC']
+            ]
+        })
+    }
+
+    const {
+        search
+    } =
+    req.query
+
+    if (search) {
+        cars = await Car.findAll({
+            where: {
+                nama: {
+                    [Op.like]: '%' + search + '%'
+                }
+            }
+        })
+    }
 
     res.render('cars', {
         title: 'halaman cars',
         layout: 'layouts/main-layout',
-        cars
+        cars,
+        date
     })
+    // res.status(200).json(cars)
 })
 
 app.get('/cars/add', (req, res) => {
@@ -65,8 +124,18 @@ app.get('/cars/add', (req, res) => {
     })
 })
 
-app.post('/cars', upload.single('foto'), async (req, res) => {
-    try {
+app.post('/cars',
+    upload.single('foto'),
+    // body('nama', 'email tidak benar!').isEmail(),
+    async (req, res) => {
+
+        const result = await validationResult(req);
+        if (!result.isEmpty()) {
+            return res.status(400).json({
+                errors: result.array()
+            });
+        }
+
         let response = await Car.create({
             nama: req.body.nama,
             harga_sewa: req.body.harga_sewa,
@@ -80,10 +149,8 @@ app.post('/cars', upload.single('foto'), async (req, res) => {
 
 
         // res.status(201).json(response)
-    } catch (err) {
-        console.log(err.message)
-    }
-})
+
+    })
 
 
 app.get('/cars/edit/:id', async (req, res) => {
@@ -95,41 +162,72 @@ app.get('/cars/edit/:id', async (req, res) => {
     })
 })
 
-app.post('/cars/update', (req, res) => {
+app.post('/cars/update', upload.single('foto'), async (req, res) => {
+
+    // const car = await Car.findOne({
+    //     id: req.params.id
+    // })
+    // if (!car) {
+    //     return res.status(404).json({
+    //         message: 'Data mobil tidak ditemukan!'
+    //     })
+    // }
+
+    // const filepath = `./public/assets/img/uploads/${car.foto}`
+    // fs.unlinkSync(filepath)
+
+
     try {
         let carId = req.body.id
-        let car = {
-            nama: req.body.nama,
-            harga_sewa: req.body.harga_sewa,
-            ukuran: req.body.ukuran,
-            foto: req.body.foto
-        }
-        let response = Car.update(car, {
+        let nama = req.body.nama
+        let harga_sewa = req.body.harga_sewa
+        let ukuran = req.body.ukuran
+        let foto = req.file.filename
+
+        await Car.update({
+            nama: nama,
+            harga_sewa: harga_sewa,
+            ukuran: ukuran,
+            foto: foto
+        }, {
             where: {
                 id: carId
             }
         })
+
         res.redirect('/cars')
-        // res.status(201).json(response)
     } catch (err) {
         console.log(err.message)
     }
 })
 app.get('/cars/delete/:id', async (req, res) => {
+    const car = await Car.findOne({
+        where: {
+            id: req.params.id
+        }
+    })
+
+    if (!car) {
+        return res.status(404).json({
+            message: 'Data mobil tidak ditemukan!'
+        })
+    }
+
     try {
-        let response = await Car.destroy({
+        const filepath = `./public/assets/img/uploads/${car.foto}`
+        fs.unlinkSync(filepath)
+
+        await Car.destroy({
             where: {
                 id: req.params.id
             }
         })
-        // res.status(201).json(response)
         res.redirect('/cars')
     } catch (err) {
         console.log(err.message)
     }
+
 })
-
-
 
 
 
@@ -137,58 +235,3 @@ app.listen(port, async (req, res) => {
     console.log(`App listening on port http://localhost:${port}/`)
     await connectToDB()
 })
-
-
-
-
-
-
-
-
-
-
-
-// import express from 'express'
-// import cors from 'cors'
-// import CarRoute from '../routes/CarRoute.js'
-// import expressLayout from 'express-ejs-layouts'
-
-// const app = express()
-// const port = 5000
-
-
-// app.set('view engine', 'ejs')
-
-// app.use('/public', express.static('public'))
-// app.use(expressLayout)
-// app.use(express.json())
-// app.use(cors())
-// app.use(CarRoute)
-
-// app.get('/', (req, res) => {
-//     res.render('index', {
-//         title: 'halaman dashboard',
-//         layout: 'layouts/main-layout'
-//     })
-// })
-
-// app.get('/cars', (req, res) => {
-//     res.render('cars', {
-//         title: 'halaman cars',
-//         layout: 'layouts/main-layout'
-//     })
-// })
-
-// app.get('/add', (req, res) => {
-//     res.render('add', {
-//         title: 'halaman add car',
-//         layout: 'layouts/main-layout'
-//     })
-// })
-
-// app.get('/edit', (req, res) => {
-//     res.render('edit', {
-//         title: 'halaman update car',
-//         layout: 'layouts/main-layout'
-//     })
-// })
